pyproject.toml
[project]
name = "ad-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.11.0",
    "einops>=0.8.1",
    "h5py>=3.15.1",
    "matplotlib>=3.10.7",
    "pyyaml>=6.0.3",
    "stable-baselines3[extra]>=2.7.0",
    "tensorboard>=2.20.0",
    "transformers>=4.57.1",
]
